# Clear Elasticsearch Data Instead of Resetting Entire Index for Specs