--- a/spec/lib/data_update_scripts/index_classified_listings_to_elasticsearch_spec.rb
+++ b/spec/lib/data_update_scripts/index_classified_listings_to_elasticsearch_spec.rb
@@ -1,9 +1,8 @@
 require "rails_helper"
 require Rails.root.join("lib/data_update_scripts/20200217215802_index_classified_listings_to_elasticsearch.rb")
 
-describe DataUpdateScripts::IndexClassifiedListingsToElasticsearch do
+describe DataUpdateScripts::IndexClassifiedListingsToElasticsearch, elasticsearch: "ClassifiedListing" do
   it "indexes classified_listings to Elasticsearch" do
-    clear_elasticsearch_data(Search::ClassifiedListing)
     classified_listing = create(:classified_listing)
     expect { classified_listing.elasticsearch_doc }.to raise_error(Search::Errors::Transport::NotFound)
     described_class.new.run