--- a/spec/lib/data_update_scripts/index_reading_list_reactions_spec.rb
+++ b/spec/lib/data_update_scripts/index_reading_list_reactions_spec.rb
@@ -1,8 +1,9 @@
 require "rails_helper"
 require Rails.root.join("lib/data_update_scripts/20200415200651_index_reading_list_reactions.rb")
 
-describe DataUpdateScripts::IndexReadingListReactions, elasticsearch: true do
+describe DataUpdateScripts::IndexReadingListReactions do
   it "indexes feed content(articles, comments, podcast episodes) to Elasticsearch" do
+    clear_elasticsearch_data(Search::Reaction)
     reactions = create_list(:reaction, 3, category: "readinglist")
     Sidekiq::Worker.clear_all