--- a/spec/lib/data_update_scripts/re_index_feed_content_to_elasticsearch_spec.rb
+++ b/spec/lib/data_update_scripts/re_index_feed_content_to_elasticsearch_spec.rb
@@ -1,9 +1,7 @@
 require "rails_helper"
 require Rails.root.join("lib/data_update_scripts/20200326145114_re_index_feed_content_to_elasticsearch.rb")
 
-describe DataUpdateScripts::ReIndexFeedContentToElasticsearch do
-  before { clear_elasticsearch_data(Search::FeedContent) }
-
+describe DataUpdateScripts::ReIndexFeedContentToElasticsearch, elasticsearch: "FeedContent" do
   after { Search::FeedContent.refresh_index }
 
   it "indexes feed content(articles, comments, podcast episodes) to Elasticsearch" do