--- a/spec/lib/data_update_scripts/re_index_users_to_elasticsearch_spec.rb
+++ b/spec/lib/data_update_scripts/re_index_users_to_elasticsearch_spec.rb
@@ -1,8 +1,9 @@
 require "rails_helper"
 require Rails.root.join("lib/data_update_scripts/20200406213152_re_index_users_to_elasticsearch.rb")
 
-describe DataUpdateScripts::ReIndexUsersToElasticsearch, elasticsearch: true do
+describe DataUpdateScripts::ReIndexUsersToElasticsearch do
   it "indexes users to Elasticsearch" do
+    clear_elasticsearch_data(Search::User)
     user = create(:user)
     expect { user.elasticsearch_doc }.to raise_error(Search::Errors::Transport::NotFound)
     sidekiq_perform_enqueued_jobs { described_class.new.run }