--- a/spec/lib/data_update_scripts/resync_elasticsearch_documents_spec.rb
+++ b/spec/lib/data_update_scripts/resync_elasticsearch_documents_spec.rb
@@ -1,7 +1,18 @@
 require "rails_helper"
 require Rails.root.join("lib/data_update_scripts/20200410152018_resync_elasticsearch_documents.rb")
 
-describe DataUpdateScripts::ResyncElasticsearchDocuments, elasticsearch: true do
+describe DataUpdateScripts::ResyncElasticsearchDocuments do
+  before do
+    clear_elasticsearch_data(Article::SEARCH_CLASS)
+    clear_elasticsearch_data(User::SEARCH_CLASS)
+    clear_elasticsearch_data(Tag::SEARCH_CLASS)
+  end
+
+  after do
+    Article::SEARCH_CLASS.refresh_index
+    User::SEARCH_CLASS.refresh_index
+  end
+
   it "indexes podcast episodes and tags to Elasticsearch" do
     tag = create(:tag)
     podcast_episode = create(:podcast_episode)