--- a/spec/lib/data_update_scripts/resync_elasticsearch_documents_spec.rb
+++ b/spec/lib/data_update_scripts/resync_elasticsearch_documents_spec.rb
@@ -1,13 +1,7 @@
 require "rails_helper"
 require Rails.root.join("lib/data_update_scripts/20200410152018_resync_elasticsearch_documents.rb")
 
-describe DataUpdateScripts::ResyncElasticsearchDocuments do
-  before do
-    clear_elasticsearch_data(Article::SEARCH_CLASS)
-    clear_elasticsearch_data(User::SEARCH_CLASS)
-    clear_elasticsearch_data(Tag::SEARCH_CLASS)
-  end
-
+describe DataUpdateScripts::ResyncElasticsearchDocuments, elasticsearch: %w[FeedContent User Tag] do
   after do
     Article::SEARCH_CLASS.refresh_index
     User::SEARCH_CLASS.refresh_index