--- a/spec/rails_helper.rb
+++ b/spec/rails_helper.rb
@@ -84,6 +84,16 @@ RSpec.configure do |config|
     Search::Cluster.recreate_indexes
   end
 
+  config.around do |ex|
+    if ex.metadata.key?(:elasticsearch)
+      klasses = Array.wrap(ex.metadata[:elasticsearch]).map do |search_class|
+        Search.const_get(search_class)
+      end
+      klasses.each { |klass| clear_elasticsearch_data(klass) }
+      ex.run
+    end
+  end
+
   config.around(:each, throttle: true) do |example|
     Rack::Attack.enabled = true
     example.run