--- a/spec/serializers/search/comment_serializer_spec.rb
+++ b/spec/serializers/search/comment_serializer_spec.rb
@@ -12,8 +12,7 @@ RSpec.describe Search::CommentSerializer do
     expect(data_hash.keys).to include(:id, :body_text, :hotness_score, :title)
   end
 
-  it "creates valid json for Elasticsearch" do
-    clear_elasticsearch_data(Search::FeedContent)
+  it "creates valid json for Elasticsearch", elasticsearch: "FeedContent" do
     data_hash = described_class.new(comment).serializable_hash.dig(:data, :attributes)
     result = Comment::SEARCH_CLASS.index(comment.id, data_hash)
     expect(result["result"]).to eq("created")