--- a/spec/serializers/search/user_serializer_spec.rb
+++ b/spec/serializers/search/user_serializer_spec.rb
@@ -8,8 +8,7 @@ RSpec.describe Search::UserSerializer do
     expect(data_hash.keys).to include(:id, :name, :path, :username, :roles)
   end
 
-  it "creates valid json for Elasticsearch" do
-    clear_elasticsearch_data(Search::User)
+  it "creates valid json for Elasticsearch", elasticsearch: "User" do
     data_hash = described_class.new(user).serializable_hash.dig(:data, :attributes)
     result = User::SEARCH_CLASS.index(user.id, data_hash)
     expect(result["result"]).to eq("created")