--- a/spec/services/search/classified_listing_spec.rb
+++ b/spec/services/search/classified_listing_spec.rb
@@ -1,10 +1,8 @@
 require "rails_helper"
 
 RSpec.describe Search::ClassifiedListing, type: :service do
-  describe "::search_documents" do
+  describe "::search_documents", elasticsearch: "ClassifiedListing" do
     let(:classified_listing) { create(:classified_listing) }
-
-    before { clear_elasticsearch_data(described_class) }
 
     it "parses classified_listing document hits from search response" do
       mock_search_response = { "hits" => { "hits" => {} } }