--- a/spec/services/search/tag_spec.rb
+++ b/spec/services/search/tag_spec.rb
@@ -1,7 +1,7 @@
 require "rails_helper"
 
 RSpec.describe Search::Tag, type: :service do
-  describe "::search_documents" do
+  describe "::search_documents", elasticsearch: "Tag" do
     let(:tag_doc_1) { { "name" => "tag1" } }
     let(:tag_doc_2) { { "name" => "tag2" } }
     let(:mock_search_response) do
@@ -14,8 +14,6 @@ RSpec.describe Search::Tag, type: :service do
         }
       }
     end
-
-    before { clear_elasticsearch_data(described_class) }
 
     it "searches with name:tag" do
       tag = create(:tag, :search_indexed, name: "tag1")