--- a/spec/workers/search/bulk_index_worker_spec.rb
+++ b/spec/workers/search/bulk_index_worker_spec.rb
@@ -1,12 +1,13 @@
 require "rails_helper"
 
-RSpec.describe Search::BulkIndexWorker, type: :worker, elasticsearch: true do
+RSpec.describe Search::BulkIndexWorker, type: :worker do
   let(:worker) { subject }
   let(:article) { create(:article) }
 
   include_examples "#enqueues_on_correct_queue", "high_priority", ["Reaction", 1]
 
   it "indexes documents for a set of given ids and object class" do
+    clear_elasticsearch_data(Search::Reaction)
     reactions = [create(:reaction, reactable: article), create(:reaction), create(:reaction)]
     Sidekiq::Worker.clear_all