--- a/spec/workers/search/index_worker_spec.rb
+++ b/spec/workers/search/index_worker_spec.rb
@@ -1,7 +1,9 @@
 require "rails_helper"
 
-RSpec.describe Search::IndexWorker, type: :worker, elasticsearch: true do
+RSpec.describe Search::IndexWorker, type: :worker do
   let(:worker) { subject }
+
+  before { clear_elasticsearch_data(Search::Tag) }
 
   include_examples "#enqueues_on_correct_queue", "high_priority", ["Tag", 1]