## What type of PR is this? (check all applicable)
- [x] Optimization

## Description
By default if we are using Elasticsearch for a spec we will reset ALL of our indexes with a very brute force approach. This is fine until you have a lot of Elasticsearch dependant tests, then it gets very sloooooow. In all of our search specs I replaced the full elasticsearch reset with a simple clear data method which uses a `delete_by_query` to ensure there is no data in the index.

So how much faster does this make the specs? Here are before and afters of all the `/search/` specs.
```
# BEFORE: Complete Reset
.........................................................................................................................

Finished in 2 minutes 40.4 seconds (files took 7.88 seconds to load)
121 examples, 0 failures

# AFTER: Only Clearing Data
.........................................................................................................................

Finished in 50.01 seconds (files took 7.73 seconds to load)
121 examples, 0 failures

````

I kept the full reset helper and use it in the `base_spec` for search bc in that spec we are deleting and creating indexes all over the place so its easier to just do a full reset for those specs. 

https://github.com/thepracticaldev/dev.to/projects/6#card-37267458

## Added tests?
- [x] yes

![alt_text](https://media1.tenor.com/images/69f6a5033c91d15323812045a40cbd42/tenor.gif?itemid=14781681)